Move TranslationConsumer -> .consumers package
